Made the feedback helper behave differently on small screens
In #6746 it was noted that on small screens the feedback form
occupies all screen space and doesn't let you read the actual
post you need to leave feedback for. This commit makes the
form appear in minimized state for narrow screens and adds
controls that allow the user to maximize and minimize the form
at his/her will.