Overhauled the small screen layout and logic
On small screens the fixed feedback form did not work very well,
this change introduces a new layout that puts the form on the
bottom of the page. The user gets notified of that form by the
floater block, which can scroll the page down and remember where
the user has been before scrolling.