Writing Helper: "<span>" + jQuery( "<div />" ).html( untrusted ).text() + "</span>" is not a safe way of stripping tags or building HTML.
Use jQuery's .text() to set the text contents of elements instead of building HTML by concatenating strings.

Strip tags "manually" with a not-generally-comprehensive regular expression. It doesn't need to be perfect because of our use of .text(). Also, the input is expected to have been run through KSES first, so this regular expression should be sufficient for this specific use case. Credit: @mdawaffe